Restore "Allow access to all URLs" option
This reverts commit c137c3118d04604b5a9266e23cda66efd153a9d9.
This also changes `<all_urls>` to `*://*/*` because the origin
permission was renamed in the MV2 to MV3 transition.

The 'Allow access to all URLs' option was removed before because the
permission was not strictly needed to access the view source option,
since that option was also shown in the crxviewer UI.

However, it does still serve the purpose of allowing users to revoke a
permission once it has been granted. This ability to revoke was also
called out in the crxviewer UI before prompting, so let's restore it.